+++ packages/pwa-kit-runtime/jest.config.js
@@ -9,6 +9,11 @@ const base = require('internal-lib-build/configs/jest/jest.config')
 
 module.exports = {
     ...base,
+    moduleNameMapper: {
+        ...base.moduleNameMapper,
+        '^@h4ad/serverless-adapter/lib/(.*)$':
+            '<rootDir>/node_modules/@h4ad/serverless-adapter/lib/$1/index.cjs'
+    },
     coverageThreshold: {
         global: {
             branches: 89,