+++ packages/pwa-kit-runtime/src/ssr/server/express.lambda.test.js
@@ -20,6 +20,7 @@ const nock = require('nock')
 const https = require('https')
 const path = require('path')
 const zlib = require('zlib')
+const {ApiGatewayV1Adapter} = require('@h4ad/serverless-adapter/lib/adapters/aws')
 
 const {X_HEADERS_TO_REMOVE_ORIGIN} = require('../../utils/ssr-proxying')
 
@@ -89,8 +90,10 @@ function createApiGatewayEvent() {
     })
 
     if (event.queryStringParameters) {
-        delete event.queryStringParameters
+        event.queryStringParameters = {}
     }
+    // aws-serverless-express added this header
+    event.headers['x-apigateway-event'] = 'apig-event'
 
     const context = AWSMockContext({
         functionName: 'SSRTest'
@@ -124,9 +127,6 @@ describe('SSRServer Lambda integration', () => {
 
     afterEach(() => {
         nock.cleanAll()
-        if (server) {
-            server.close()
-        }
     })
 
     const fakeBinaryPayload = crypto.randomBytes(16)
@@ -323,9 +323,15 @@ describe('SSRServer Lambda integration', () => {
                 path: testCase.path,
                 body: undefined
             })
+            // aws-serverless-express added this header
+            event.headers['x-apigateway-event'] = 'apig-event'
 
+            // Check to make sure the adapter can handle the event
+            expect(new ApiGatewayV1Adapter().canHandle(event)).toBe(true)
+
+            // AWS API Gateway adapter expects queryStringParameters key to exist within the event
             if (event.queryStringParameters) {
-                delete event.queryStringParameters
+                event.queryStringParameters = {}
             }
 
             // Add a fake X-Amz-Cf-Id header
@@ -357,6 +363,7 @@ describe('SSRServer Lambda integration', () => {
 
                         // We expect a context property to have been set false
                         expect(context.callbackWaitsForEmptyEventLoop).toBe(false)
+                        expect(response.headers['date']).toBeDefined()
 
                         // Check the response
                         testCase.validate(response)
@@ -401,7 +408,7 @@ describe('SSRServer Lambda integration', () => {
         })
 
         if (event.queryStringParameters) {
-            delete event.queryStringParameters
+            event.queryStringParameters = {}
         }
 
         const context = AWSMockContext({